factor logging stubs out for easier testing
