+++ Makefile.am
@@ -200,7 +200,9 @@ integration: all py-integration
 tests_art_t_CPPFLAGS = $(THIRDPARTY_CPPFLAGS)
 tests_art_t_LDADD = $(ART_LIB) $(TAP_LIB)
 tests_art_t_SOURCES = \
-	tests/art_test.c
+	tests/art_test.c \
+	tests/log_stub.c \
+	log.c
 
 tests_argv_t_CPPFLAGS = $(THIRDPARTY_CPPFLAGS)
 tests_argv_t_LDADD = $(JSON_LIB) $(TAP_LIB)
@@ -212,6 +214,7 @@ tests_ignore_t_CPPFLAGS = $(THIRDPARTY_CPPFLAGS)
 tests_ignore_t_LDADD = $(ART_LIB) $(TAP_LIB) $(JSON_LIB)
 tests_ignore_t_SOURCES = \
 	tests/ignore_test.c \
+	tests/log_stub.c \
 	watcher/helper.c \
 	hash.c \
 	ht.c \
